primaverum/pyproject.toml
[project]
name = "primaverum"
version = "0.1.0"
description = "primaverum using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[litellm,tools]==1.14.1",
    "litellm>=1.83.0",
    "nvidia-riva-client>=2.25.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.56.0",
]

[project.scripts]
primaverum = "primaverum.main:run"
run_crew = "primaverum.main:run"
train = "primaverum.main:train"
replay = "primaverum.main:replay"
test = "primaverum.main:test"
run_with_trigger = "primaverum.main:run_with_trigger"
run_gui = "primaverum.main:run_gui"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
